fix error propagation in other video workers
fix circular imports by refactoring and moving utils to separate files

able to basic compilation with human feedback nodes

create tool_nodes used in human_feedback
