+++ jockey/cli.py
@@ -2,13 +2,14 @@ import uuid
 import os
 import subprocess
 from rich.console import Console
-from jockey.util import parse_langchain_events_terminal, get_langgraph_errors
+from jockey.util import parse_langchain_events_terminal
+from jockey.stirrups.errors import create_interrupt_event, create_langgraph_error_event, create_jockey_error_event, get_langgraph_errors
 from langchain_core.messages import HumanMessage
 from jockey.app import jockey
 from jockey.stirrups.errors import JockeyError
 import asyncio
-from jockey.util import create_interrupt_event, create_langgraph_error_event, create_jockey_error_event
 import sys
+from jockey.jockey_graph import JockeyState, Jockey
 
 
 async def process_single_message(message: str):
@@ -42,6 +43,7 @@ async def run_jockey_terminal():
     """Quickstart function to create a Jockey instance in the terminal for easy dev work."""
     console = Console()
     session_id = uuid.uuid4()
+    thread = {"configurable": {"thread_id": session_id}}
 
     try:
         console.print()
@@ -54,11 +56,40 @@ async def run_jockey_terminal():
 
         try:
             last_event = None
-            stream = jockey.astream_events(jockey_input, {"configurable": {"thread_id": session_id}}, version="v2")
+            stream = jockey.astream_events(jockey_input, thread, version="v2")
 
             async for event in stream:
                 last_event = event
                 await parse_langchain_events_terminal(event)
+
+            # human feedback section
+            # get user input
+            try:
+                user_input = console.input("[green]👤 Feedback: ")
+            except KeyboardInterrupt:
+                user_input = "no feedback"
+                console.print("\nExiting Jockey terminal...")
+                sys.exit(0)
+
+            # update jockey state with feedback
+            jockey_state_human_feedback = JockeyState(
+                chat_history=last_event.get("chat_history", []),
+                next_worker=last_event.get("next_worker", None),
+                made_plan=last_event.get("made_plan", False),
+                active_plan=last_event.get("active_plan", None),
+                feedback=user_input,
+            )
+
+            jockey_graph.update_state(thread, jockey_state_human_feedback, as_node="human_feedback")
+
+            # debug
+            print("--State after update--")
+            print(jockey_graph.get_state(thread))
+            jockey_graph.get_state(thread).next
+
+            # proceed
+            for event in jockey_graph.stream(None, thread, stream_mode="values"):
+                print(event)
 
         except asyncio.CancelledError:
             console.print("\nOperation interrupted")
