+++ jockey/stirrups/video_text_generation.py
@@ -6,10 +6,10 @@ from langchain.pydantic_v1 import BaseModel, Field
 from langchain.tools import tool
 from typing import Dict, List, Union
 from enum import Enum
-from jockey.util import create_jockey_error_event, get_video_metadata, parse_langchain_events_terminal
+from jockey.video_utils import get_video_metadata
 from jockey.prompts import DEFAULT_VIDEO_TEXT_GENERATION_FILE_PATH
 from jockey.stirrups.stirrup import Stirrup
-from jockey.stirrups.errors import ErrorType, JockeyError, NodeType, WorkerFunction
+from jockey.stirrups.errors import ErrorType, JockeyError, NodeType, WorkerFunction, create_jockey_error_event
 
 
 TL_BASE_URL = "https://api.twelvelabs.io/v1.2/"
@@ -85,8 +85,6 @@ async def gist_text_generation(video_id: str, index_id: str, endpoint_options: L
             function_name=WorkerFunction.GIST_TEXT_GENERATION,
             details=f"Error: {str(error)}",
         )
-        jockey_error_event = create_jockey_error_event(error=jockey_error)
-        await parse_langchain_events_terminal(jockey_error_event)
         raise jockey_error
 
 
@@ -115,13 +113,11 @@ async def summarize_text_generation(video_id: str, index_id: str, endpoint_optio
             function_name=WorkerFunction.SUMMARIZE_TEXT_GENERATION,
             details=f"Error: {str(error)}",
         )
-        jockey_error_event = create_jockey_error_event(error=jockey_error)
-        await parse_langchain_events_terminal(jockey_error_event)
         raise jockey_error
 
 
 @tool("freeform-text-generation", args_schema=PegasusFreeformInput)
-async def free_text_generation(video_id: str, index_id: str, prompt: str) -> Dict:
+async def freeform_text_generation(video_id: str, index_id: str, prompt: str) -> Dict:
     """Generate any type of text output for a single video.
     Useful for answering specific questions, understanding fine grained details, and anything else that doesn't fall neatly into the other tools."""
     try:
@@ -141,17 +137,15 @@ async def free_text_generation(video_id: str, index_id: str, prompt: str) -> Dic
         jockey_error = JockeyError.create(
             node=NodeType.WORKER,
             error_type=ErrorType.TEXT_GENERATION,
-            function_name=WorkerFunction.FREE_TEXT_GENERATION,
+            function_name=WorkerFunction.FREEFORM_TEXT_GENERATION,
             details=f"Error: {str(error)}",
         )
-        jockey_error_event = create_jockey_error_event(error=jockey_error)
-        await parse_langchain_events_terminal(jockey_error_event)
         raise jockey_error
 
 
 # Construct a valid worker for a Jockey instance.
 video_text_generation_worker_config = {
-    "tools": [gist_text_generation, summarize_text_generation, free_text_generation],
+    "tools": [gist_text_generation, summarize_text_generation, freeform_text_generation],
     "worker_prompt_file_path": DEFAULT_VIDEO_TEXT_GENERATION_FILE_PATH,
     "worker_name": "video-text-generation",
 }
